admin and user role and permission for delete and manage user

+++ app/Http/Controllers/ProjectController.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\PermissionEnum;
 use App\Models\Client;
 use App\Models\Project;
 use App\Http\Requests\StoreProjectRequest;
 use App\Http\Requests\UpdateProjectRequest;
 use App\Models\User;
-use Illuminate\Http\Client\Request;
-use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Gate;
+
 
 class ProjectController extends Controller
 {
@@ -64,6 +65,8 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+        Gate::authorize(PermissionEnum::DELETE_PROJECTS->value);
+
         $project->delete();
         return redirect()->route('projects.index');
     }
